+++ community/server/src/main/java/org/neo4j/server/rest/management/console/ConsoleSessionFactory.java
@@ -17,14 +17,13 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.server.webadmin.console;
+package org.neo4j.server.rest.management.console;
 
-import org.neo4j.helpers.collection.Pair;
+import org.neo4j.logging.LogProvider;
+import org.neo4j.server.database.Database;
 
-public interface ScriptSession
+public interface ConsoleSessionFactory
 {
-    /**
-     * @return {@link Pair} of (result,next prompt).
-     */
-    Pair<String, String> evaluate( String script );
+    ScriptSession createSession( String engineName, Database database, LogProvider logProvider );
+    Iterable<String> supportedEngines();
 }
