+++ community/server/src/main/java/org/neo4j/server/rest/management/console/ScriptSession.java
@@ -17,17 +17,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.server.webadmin.console;
+package org.neo4j.server.rest.management.console;
 
-import javax.servlet.http.HttpServletRequest;
+import org.neo4j.helpers.collection.Pair;
 
-import org.neo4j.logging.LogProvider;
-import org.neo4j.server.database.CypherExecutor;
-import org.neo4j.server.database.Database;
-
-public interface ConsoleSessionCreator
+public interface ScriptSession
 {
-    String name();
-
-    ScriptSession newSession( Database database, CypherExecutor cypherExecutor, HttpServletRequest request, LogProvider logProvider );
+    /**
+     * @return {@link Pair} of (result,next prompt).
+     */
+    Pair<String, String> evaluate( String script );
 }
