+++ community/server/src/main/resources/META-INF/services/org.neo4j.server.rest.management.console.ConsoleSessionCreator
