+++ community/server/src/test/java/org/neo4j/server/webadmin/rest/ConsoleServiceDocTest.java
@@ -33,8 +33,8 @@ import org.neo4j.server.database.Database;
 import org.neo4j.server.rest.management.console.ConsoleService;
 import org.neo4j.server.rest.repr.OutputFormat;
 import org.neo4j.server.rest.repr.formats.JsonFormat;
-import org.neo4j.server.webadmin.console.ConsoleSessionFactory;
-import org.neo4j.server.webadmin.console.ScriptSession;
+import org.neo4j.server.rest.management.console.ConsoleSessionFactory;
+import org.neo4j.server.rest.management.console.ScriptSession;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
