+++ community/server/src/test/java/org/neo4j/server/webadmin/rest/Neo4jShellConsoleSessionDocTest.java
@@ -41,8 +41,8 @@ import org.neo4j.server.rest.management.console.ConsoleService;
 import org.neo4j.server.rest.management.console.ShellSession;
 import org.neo4j.server.rest.repr.OutputFormat;
 import org.neo4j.server.rest.repr.formats.JsonFormat;
-import org.neo4j.server.webadmin.console.ConsoleSessionFactory;
-import org.neo4j.server.webadmin.console.ScriptSession;
+import org.neo4j.server.rest.management.console.ConsoleSessionFactory;
+import org.neo4j.server.rest.management.console.ScriptSession;
 import org.neo4j.shell.ShellSettings;
 import org.neo4j.test.TestGraphDatabaseFactory;
